Color channels upscaled by bicubic interpolation
For color images, we previously applied our RAISR algorithm only to the luminance channel, whereas in the Google's RAISR implementation and by others, the color channels are upsampled by bicubic interpolation. We can also use bicubic interpolation, but this will not affect our reported PSNR results as it is dependent on luminance only. Based on comments in issue 1, in this commit, we are switching to applying bicubic interpolation on the chromatic channels.